Projects/chat_with_pdfs/requirements.txt
langchain
openai
chromadb
PyPDF2
streamlit
python-dotenv
tqdm
langchain-community
pypdf
python-dotenv
tiktoken
